Spiking an idea for formatted notes
Right now the "notes" for a slide are intended to be something short and
sweet, especially given that they're also allowed to be shown on the
normal view. It's specifically treated as a simple p.notes tag and the
contained *text* is grabbed by JS and inserted in the right place.

The template presentation supplied to us tried extending notes by having
a bunch of lines starting with ".notes" instead of just one.  That just
results in a messy paragraph with ".notes" appearing in it a bunch, but
it's an interesting way to show what all the notes are.

An overview of changes in this spike:

 - Take the p.notes content, strip out all the extra ".notes" appearing
   in there, and process that as Markdown.
 - Take that Markdown result and make it the content of a div.notes tag
   (replacing the p.notes, since p is not a block-level tag)
 - Change the JS to look for div.notes instead of p.notes
 - Put the HTML content of div.notes in the presenter notes area (still
   using just the text for the regular view slide-footer notes)
 - Update the presenter CSS for the notes area to show lists nicely and
   handle vertical overflow (with scrolling)