Spiking an idea for analytics
It has to be JS-based because nothing happens on the server once all the
slides are loaded. On the other hand, getting JS to output to a file is
a huge pain. Solution: use the built-in JS events (like showoff:show) to
contact the server and let it know you're on a new slide.

  - Creating analytics log file on server start
  - Added 'ping' action
    - logs a "ping" to the analytics file
    - records client IP (for tracking multiple viewers)
    - records slide (parameter included in ping request)
  - piggybacking on showoff:show event
    - sends request to the 'ping' action
    - includes slide name (taken from the event target's "ref" attr)