Adding special content functionality
This allows the user to specify sections like .notes and .handouts that
are handled in special ways. For example, the .handouts section is
styled so that it only appears when the slideshow is printed and the
.notes appear in the presenter view notes box. This should also process
markup so that the contents of these sections can be formatted.

Starting a test for "special content"

Right now just working on the conversion that showoff already does.

Also pointing out that position within the slide doesn't matter (within
reason), and it's okay to split longer notes into multiple lines.

Handling longer "special content"

Now it's possible to have notes that not only span multiple lines, but
also include blank lines and indented text. For these things (notably
the blank lines), it's necessary to repeat the marker for every line,
and that marker is then taken out.

Moved the content alteration from operating on the final (entire file)
to doing it on a slide-by-slide basis. It was important for the way I
was handling the notes, but more than that I think it's proper to treat
each slide on its own, even if many of them happen to be defined in a
single file.

Also making sure there's no problem if a slide has no notes (either
because the line doesn't exist or there is no note text).

Better updating of "special content"

Using Nokogiri fully, not half-way. The content can be set without silly
regex tricks, and DocumentFragment is a great way to keep the resulting
HTML from being wrapped with extra stuff.

Refactoring "special content" test.

Putting in some useful helper methods and variables, making the tests
all look the same.

Adding formatting to "special content"

Changing the plain p.notes to div.notes with markdown-formatted content.

Needed to make quite a few changes to the test, mostly in the way of
convenience or helper methods for getting or normalizing the notes. One
of my least-favorite things is trying to compare two bits of HTML to
see if they're equivalent. Even with Nokogiri, this isn't so much fun.

Caught a repeated test name, which was keeping a test from being run
(redefining the test instead of having both defined).

JS/CSS changes for formatted notes

Now that the notes are HTML and not text, the JS that puts the notes in
the right place (for viewers or the presenter) needs to be changed.
Still using just the text for the slide footer the viewers see, but the
presenter now gets the entire notes HTML.

The presenter notes also need to be expanded a bit, allow scrolling for
longer notes, and show lists correctly.

Allowing formatting for other "special content"

Special content is wanted for three things: presenter notes (originally
part of showoff, but now extended), handouts, and exercises. This takes
the code already there for the notes, moves that to a reusable method,
and also applies it to handouts and exercises.

Hiding all "special content"

Notes, handout, and exercise content are all hidden by default.

Remove duplicate definition for getCurrentnotes

There was possible a merge problem? (ask Yossef), but this 2nd definition was screwing
things up.

Properly parses styles in .notes/.handouts

Previous to this patch, showoff would throw away all html after
processing markdown. So only text gets included in .notes/.handouts.

I don't know why the markdown processor needs to be invoked twice,
but that's a battle for another day.