Add a dropdown selector for alternate styles
Starting clean style picker, using alt stylesheets

Adding all file stylesheets as alternates, not regular.

Adding stylemenu to normal view.

Pretty much just like the table of contents/navmenu.

A mass of style-choice-related JS

The style menu is modeled on the table of contents, but more complicated
in some cases (getting the menu choices from the links) and simpler in
others (the menu will always be a single level, no nesting).

There's also what's to be done when a choice is selected, which is
"deactivate" (make alternate) all the stylesheet links except for the
chosen one, which is "activated" (made to be just a regular
'stylesheet').

Cleaning up style menu visual placement.

Keeping the style menu from pushing the slide down the page when visible.

Adding style menu hotkey

Now you can press 's' to bring up the menu and pick style to your
heart's content.

Changing the style in the presenter view as well.

Now a change to the style in the normal view also affects the presenter
view.

Tying the presenter/regular view styles together.

Now you can get the style menu on the presenter view as well, and a
change to the style on either view (presenter or normal) will be
communicated to the other. The views should now always stay in sync.

The 'prop' argument is for 'propagate', to let the views communicate
back and forth the right amount, without turning this into a massive
stack-busting call loop.

Adding way to set default style.

Now that all the custom stylesheets are loaded as alternates, the
presentation will be unstyled until a style is chosen from the menu.
This provides a way to specific (in the presentation config file) what
the style should be to start.

Adding default-style predicate for convenience

Rather than check each style manually, this method will tell you whether
a given style is the default.

Isn't there a `deny` or `reject` counterpoint to `assert`? If there is,
I can't remember or find it.

Setting default stylesheet on presentation start.

Now the one stylesheet matching the style specified in the config file
will be marked as a normal stylesheet instead of an alternate.

Chrome doesn't understand the alternate stylesheet tag