First stab at a follow mode
Students following along can press 'g' to enable follow mode. When
follow mode is on, their slide deck will advance to keep up with the
presenter. I am dubious of the actual value of this feature, but there
seemed to be a lot of interest in the functionality.

I am on a plane right now, so I cannot look up details. There are a
couple items that I need to look up and then this feature will be
complete.

TODO:
* The class variable @@current will print to the console, but I cannot
get it to return a valid value to the client.
* I forget the implementation details of the ajax xmlhttprequest. I just
need to get the response body, so minor detail.

Finished follow mode