Make follow mode more reliable
Always returning the current page instead of only returning it when a
client is pinging seems to make the process more reliable. I'm not
entirely sure why this is, but let's roll with it for now and observe.

It's almost looking like being called with a null value might be
confusing my page updater.